Designed UploadSourceCode form that contains three buttons ,1 text area ,1 combo box and 3 text labels
The buttons conclude "Back" that returns to main menu,"Start Practice "that starts the practice immediately  and "Save" that saves the entered code. 
The combo box is called "Language" that contains the supported languages for training
The text area's called "Code" that is supposed to be the place where the user enters the code.
The labels are "Upload Source Code","Language" and "Code".