Designed PracticeForm that contains 2 buttons,2 text areas and 8 labels
The buttons are "Start" that will start the timer and "EXIT" that exits to the main menu.
The text areas are "Original"that contains the original code and "User Input"that is supposed to be the area to let the user start coding .
The labels are "Practice!","Language","Language chosen","Time","Timer","OriginalCode","UserInput" and "Caution Message".